gate rules added for comment-delete

+++ app/Http/Controllers/CommentController.php
@@ -4,14 +4,21 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Comment;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function create(Request $request)
     {
         $comment = new Comment;
         $comment->content = $request->content;
         $comment->article_id = $request->article_id;
+        $comment->user_id = auth()->user()->id;
         $comment->save();
 
         return back();
@@ -20,8 +27,11 @@ class CommentController extends Controller
     public function delete($id)
     {
         $comment = Comment::find($id);
-        $comment->delete();
-
-        return back();
+        if( Gate::allows('comment-delete', $comment) ) {
+            $comment->delete();
+            return back();
+        } else {
+            return back()->with('error', 'Unauthorize');
+        }
     }
 }
